+++ third-springboot/src/main/java/org/sopt/www/firstspringboot/exception/model/BusinessException.java
@@ -1,4 +1,4 @@
-package org.sopt.www.firstspringboot.exception;
+package org.sopt.www.firstspringboot.exception.model;
 
 public class BusinessException extends RuntimeException {
     public BusinessException(String message) {
