+++ third-springboot/src/main/resources/application.yml
@@ -1,38 +1,38 @@
 # H2
-spring:
-  datasource:
-    driver-class-name: org.h2.Driver
-    url: jdbc:h2:tcp://localhost/~/test
-    username: sa
-    password:
-
-  jpa:
-    defer-datasource-initialization: true
-    show-sql: true
-    hibernate:
-      ddl-auto: create
-    properties:
-      hibernate:
-        format_sql: true
-        show_sql: true
-
-  sql:
-    init:
-      mode: always
-
-# DataGrip
 #spring:
 #  datasource:
-#    driver-class-name: com.mysql.cj.jdbc.Driver
-#    url: jdbc:mysql://localhost:3306
-#    username: root
-#    password: rina030429
+#    driver-class-name: org.h2.Driver
+#    url: jdbc:h2:tcp://localhost/~/test
+#    username: sa
+#    password:
 #
+#  jpa:
+#    defer-datasource-initialization: true
+#    show-sql: true
+#    hibernate:
+#      ddl-auto: create
+#    properties:
+#      hibernate:
+#        format_sql: true
+#        show_sql: true
+#
+#  sql:
+#    init:
+#      mode: always
+
+# DataGrip
+spring:
+  datasource:
+    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: jdbc:mysql://sopt-db.cfaclnkmujol.ap-northeast-2.rds.amazonaws.com:3306/soptdb
+    username: admin
+    password: rina030429
+
 #  config:
 #    activate:
 #      on-profile: dev
-#
-#  jpa:
-#    database-platform: org.hibernate.dialect.MySQL8Dialect
-#    hibernate:
-#      ddl-auto: create
+
+  jpa:
+    database-platform: org.hibernate.dialect.MySQL8Dialect
+    hibernate:
+      ddl-auto: create
